refactor(cli): flatten goal command module
Move the Goal command slice to the neutral command root and update its internal export without changing public CLI or API behavior.
